Implementada verificação para impedir o cadastro de dispositivos repetidos

+++ src/contexts/MicrocontrollersContext.jsx
@@ -1,35 +1,38 @@
-import {createContext, useEffect, useState} from 'react'
+import { createContext, useEffect, useState } from 'react'
 
 import AsyncStorage from '@react-native-async-storage/async-storage'
 
 const MicrocontrollersContext = createContext({})
 
-function MicrocontrollersContextProvider({children}){
+function MicrocontrollersContextProvider({ children }) {
     const [storedDevices, setStoredDevices] = useState([])
 
-    async function getSavedDevices(){
+    async function getSavedDevices() {
         const jsonCurrentSavedDevices = await AsyncStorage.getItem('@Microcontrollers')
         const currentSavedDevices = JSON.parse(jsonCurrentSavedDevices)
-        if(currentSavedDevices === null){
+        if (currentSavedDevices === null) {
             return []
         }
         return currentSavedDevices
 
     }
 
-    function loadDevices(){
+    function loadDevices() {
         getSavedDevices().then(devices => {
             setStoredDevices(devices)
         })
     }
 
-    async function saveNewDevice(databaseId, address, id, name){
+    async function saveNewDevice(databaseId, address, id, name) {
         const savedDevices = await getSavedDevices()
-        const newDevice = {address, id, name, databaseId}
-        await AsyncStorage.setItem('@Microcontrollers', JSON.stringify([...savedDevices, newDevice]))
-        loadDevices()
+        const alreadyInUseDevices = storedDevices.filter(device => device.id === id || device.name === name)
+        if (!(alreadyInUseDevices.length > 0)) {
+            const newDevice = { address, id, name, databaseId }
+            await AsyncStorage.setItem('@Microcontrollers', JSON.stringify([...savedDevices, newDevice]))
+            loadDevices()
+        }
     }
-    async function removeDevice(address){
+    async function removeDevice(address) {
         const filteredDevices = storedDevices.filter(device => device.address !== address)
         AsyncStorage.setItem('@Microcontrollers', JSON.stringify(filteredDevices))
         loadDevices()
@@ -39,16 +42,17 @@ function MicrocontrollersContextProvider({children}){
         console.log(`Stored devices: ${JSON.stringify(storedDevices)}`)
     }, [storedDevices])
 
-    useEffect(()=>{
+    useEffect(() => {
         loadDevices()
         // AsyncStorage.clear()
     }, [])
 
     return (
-        <MicrocontrollersContext.Provider value={{saveNewDevice, removeDevice, storedDevices}}>
+        <MicrocontrollersContext.Provider value={{ saveNewDevice, removeDevice, storedDevices }}>
             {children}
         </MicrocontrollersContext.Provider>
     )
 }
 
-export {MicrocontrollersContextProvider, MicrocontrollersContext}
+export { MicrocontrollersContextProvider, MicrocontrollersContext }
+
